issue open events added for project, cycle, module, archives, notification
