+++ web/components/dashboard/widgets/issue-panels/issues-list.tsx
@@ -14,11 +14,12 @@ import {
   CreatedUpcomingIssueListItem,
   IssueListItemProps,
 } from "@/components/dashboard/widgets";
-// ui
+// constants
+import { E_DASHBOARD, ISSUE_OPENED } from "constants/event-tracker";
 // helpers
 import { cn } from "@/helpers/common.helper";
 import { getRedirectionFilters } from "@/helpers/dashboard.helper";
-import { useIssueDetail } from "@/hooks/store";
+import { useIssueDetail, useEventTracker } from "@/hooks/store";
 // types
 
 export type WidgetIssuesListProps = {
@@ -33,9 +34,19 @@ export const WidgetIssuesList: React.FC<WidgetIssuesListProps> = (props) => {
   const { isLoading, tab, type, widgetStats, workspaceSlug } = props;
   // store hooks
   const { setPeekIssue } = useIssueDetail();
+  const { captureEvent } = useEventTracker();
 
-  const handleIssuePeekOverview = (issue: TIssue) =>
+  const handleIssuePeekOverview = (issue: TIssue) => {
     setPeekIssue({ workspaceSlug, projectId: issue.project_id, issueId: issue.id });
+    captureEvent(ISSUE_OPENED, {
+      element: E_DASHBOARD,
+      element_id: tab,
+      mode: "peek",
+      filters: {
+        target_date: issue.target_date,
+      },
+    });
+  };
 
   const filterParams = getRedirectionFilters(tab);
