+++ web/components/issues/issue-layouts/roots/cycle-layout-root.tsx
@@ -133,7 +133,7 @@ export const CycleLayoutRoot: React.FC = observer(() => {
               ) : null}
             </div>
             {/* peek overview */}
-            <IssuePeekOverview />
+            <IssuePeekOverview issuesFilter={issuesFilter.issueFilters} />
           </Fragment>
         )}
       </div>
