+++ web/components/issues/issue-layouts/roots/project-layout-root.tsx
@@ -86,7 +86,7 @@ export const ProjectLayoutRoot: FC = observer(() => {
           </div>
 
           {/* peek overview */}
-          <IssuePeekOverview />
+          <IssuePeekOverview issuesFilter={issuesFilter.issueFilters} />
         </Fragment>
       )}
     </div>
