+++ web/components/issues/peek-overview/header.tsx
@@ -1,6 +1,7 @@
 import { FC } from "react";
 import { observer } from "mobx-react";
 import Link from "next/link";
+import { useRouter } from "next/router";
 import { ArchiveRestoreIcon, Link2, MoveDiagonal, MoveRight, Trash2 } from "lucide-react";
 // ui
 import {
@@ -16,11 +17,14 @@ import {
 // components
 import { IssueSubscription, IssueUpdateStatus } from "@/components/issues";
 import { STATE_GROUPS } from "@/constants/state";
+// constants
+import { ISSUE_OPENED } from "constants/event-tracker";
 // helpers
 import { cn } from "@/helpers/common.helper";
+import { getElementIdFromPath } from "@/helpers/event-tracker.helper";
 import { copyUrlToClipboard } from "@/helpers/string.helper";
 // store hooks
-import { useIssueDetail, useProjectState, useUser } from "@/hooks/store";
+import { useIssueDetail, useProjectState, useUser, useEventTracker } from "@/hooks/store";
 // hooks
 import { usePlatformOS } from "@/hooks/use-platform-os";
 export type TPeekModes = "side-peek" | "modal" | "full-screen";
@@ -73,6 +77,8 @@ export const IssuePeekOverviewHeader: FC<PeekOverviewHeaderProps> = observer((pr
     handleRestoreIssue,
     isSubmitting,
   } = props;
+  // router
+  const router = useRouter();
   // store hooks
   const { data: currentUser } = useUser();
   const {
@@ -80,6 +86,7 @@ export const IssuePeekOverviewHeader: FC<PeekOverviewHeaderProps> = observer((pr
   } = useIssueDetail();
   const { getStateById } = useProjectState();
   const { isMobile } = usePlatformOS();
+  const { captureEvent } = useEventTracker();
   // derived values
   const issueDetails = getIssueById(issueId);
   const stateDetails = issueDetails ? getStateById(issueDetails?.state_id) : undefined;
@@ -118,7 +125,17 @@ export const IssuePeekOverviewHeader: FC<PeekOverviewHeaderProps> = observer((pr
         </Tooltip>
 
         <Tooltip tooltipContent="Open issue in full screen" isMobile={isMobile}>
-          <Link href={`/${issueLink}`} onClick={() => removeRoutePeekId()}>
+          <Link
+            href={`/${issueLink}`}
+            onClick={() => {
+              removeRoutePeekId();
+              captureEvent(ISSUE_OPENED, {
+                element: "peek",
+                elementId: getElementIdFromPath(router.asPath),
+                mode: "detail",
+              });
+            }}
+          >
             <MoveDiagonal className="h-4 w-4 text-custom-text-300 hover:text-custom-text-200" />
           </Link>
         </Tooltip>
