+++ web/components/notifications/notification-popover.tsx
@@ -63,12 +63,12 @@ export const NotificationPopover = observer(() => {
   const currentTabEmptyState = snoozed
     ? EmptyStateType.NOTIFICATION_SNOOZED_EMPTY_STATE
     : archived
-    ? EmptyStateType.NOTIFICATION_ARCHIVED_EMPTY_STATE
-    : selectedTab === "created"
-    ? EmptyStateType.NOTIFICATION_CREATED_EMPTY_STATE
-    : selectedTab === "watching"
-    ? EmptyStateType.NOTIFICATION_SUBSCRIBED_EMPTY_STATE
-    : EmptyStateType.NOTIFICATION_MY_ISSUE_EMPTY_STATE;
+      ? EmptyStateType.NOTIFICATION_ARCHIVED_EMPTY_STATE
+      : selectedTab === "created"
+        ? EmptyStateType.NOTIFICATION_CREATED_EMPTY_STATE
+        : selectedTab === "watching"
+          ? EmptyStateType.NOTIFICATION_SUBSCRIBED_EMPTY_STATE
+          : EmptyStateType.NOTIFICATION_MY_ISSUE_EMPTY_STATE;
 
   return (
     <>
@@ -154,6 +154,8 @@ export const NotificationPopover = observer(() => {
                           selectedTab={selectedTab}
                           key={notification.id}
                           isSnoozedTabOpen={snoozed}
+                          isArchivedTabOpen={archived}
+                          isUnreadTabOpen={readNotification}
                           closePopover={() => setIsActive(false)}
                           notification={notification}
                           markNotificationArchivedStatus={markNotificationArchivedStatus}
